+++ ibeis/expt/experiment_printres.py
@@ -297,7 +297,6 @@ def print_results(ibs, testres, **kwargs):
             #frac_list = (nLessX_dict[int(X)] / cfgx2_nQuery)[:, None]
             #print('cfgx2_nQuery = %r' % (cfgx2_nQuery,))
             #print('frac_list = %r' % (frac_list,))
-            #print('Pairwise Difference: ' + str(ut.safe_pdist(frac_list, metric=ut.absdiff)))
             for cfgx in sortx:
                 nLessX_ = nLessX_dict[int(X)][cfgx]
                 rankstr = rankscore_str(X, nLessX_, cfgx2_nQuery[cfgx], withlbl=False)